+++ FoodTruckBe/src/main/java/com/toss/foodtruck/domain/auth/service/AuthService.java
@@ -3,8 +3,8 @@ package com.toss.foodtruck.domain.auth.service;
 import com.toss.foodtruck.domain.auth.dto.LoginRequestDto;
 import com.toss.foodtruck.domain.auth.dto.TokenResponseDto;
 import com.toss.foodtruck.domain.auth.dto.TossResponseDto;
-import com.toss.foodtruck.domain.entity.Member;
-import com.toss.foodtruck.domain.repository.MemberRepository;
+import com.toss.foodtruck.domain.member.entity.Member;
+import com.toss.foodtruck.domain.member.repository.MemberRepository;
 import com.toss.foodtruck.global.enums.Role;
 import com.toss.foodtruck.global.jwt.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
